+++ lib/pagy/extras/searchkick.rb
@@ -2,9 +2,18 @@
 # encoding: utf-8
 # frozen_string_literal: true
 
-require 'pagy/extras/pagy_search'
-
 class Pagy
+
+  module Searchkick
+    # returns an array used to delay the call of #search
+    # after the pagination variables are merged to the options
+    # it also pushes to the same array an eventually called method and arguments
+    def pagy_search(term = "*", **options, &block)
+      [self, term, options, block].tap do |args|
+        args.define_singleton_method(:method_missing){|*a| args += a}
+      end
+    end
+  end
 
   # create a Pagy object from a Searchkick::Results object
   def self.new_from_searchkick(results, vars={})
@@ -19,12 +28,12 @@ class Pagy
 
     # Return Pagy object and results
     def pagy_searchkick(pagy_search_args, vars={})
-      model, search_args, block, *called = pagy_search_args
-      vars                       = pagy_searchkick_get_vars(nil, vars)
-      search_args[-1][:per_page] = vars[:items]
-      search_args[-1][:page]     = vars[:page]
-      results                    = model.search(*search_args, &block)
-      vars[:count]               = results.total_count
+      model, term, options, block, *called = pagy_search_args
+      vars               = pagy_searchkick_get_vars(nil, vars)
+      options[:per_page] = vars[:items]
+      options[:page]     = vars[:page]
+      results            = model.search(term, **options, &block)
+      vars[:count]       = results.total_count
       pagy = Pagy.new(vars)
       # with :last_page overflow we need to re-run the method in order to get the hits
       if defined?(Pagy::Overflow) && pagy.overflow? && pagy.vars[:overflow] == :last_page