+++ test/pagy/extras/searchkick_test.rb
@@ -7,7 +7,7 @@ require 'pagy/extras/overflow'
 
 SimpleCov.command_name 'elasticsearch' if ENV['RUN_SIMPLECOV'] == 'true'
 
-describe Pagy::Search do
+describe Pagy::Searchkick do
 
   describe '#pagy_search' do
 
@@ -16,29 +16,29 @@ describe Pagy::Search do
     end
 
     it 'returns class and arguments' do
-      _(MockSearchkick::Model.pagy_search('a', b:2)).must_equal [MockSearchkick::Model, ['a', {b: 2}], nil]
+      _(MockSearchkick::Model.pagy_search('a', b:2)).must_equal [MockSearchkick::Model, 'a', {b: 2}, nil]
       args  = MockSearchkick::Model.pagy_search('a', b:2){|a| a*2}
       block = args[-1]
-      _(args).must_equal [MockSearchkick::Model, ['a', {b: 2}], block]
+      _(args).must_equal [MockSearchkick::Model, 'a', {b: 2}, block]
     end
 
     it 'allows the term argument to be optional' do
-      _(MockSearchkick::Model.pagy_search(b:2)).must_equal [MockSearchkick::Model, [{b: 2}], nil]
+      _(MockSearchkick::Model.pagy_search(b:2)).must_equal [MockSearchkick::Model, '*', {b: 2}, nil]
       args  = MockSearchkick::Model.pagy_search(b:2){|a| a*2}
       block = args[-1]
-      _(args).must_equal [MockSearchkick::Model, [{b: 2}], block]
+      _(args).must_equal [MockSearchkick::Model, '*', {b: 2}, block]
     end
 
     it 'adds an empty option hash' do
-      _(MockSearchkick::Model.pagy_search('a')).must_equal [MockSearchkick::Model, ['a', {}], nil]
+      _(MockSearchkick::Model.pagy_search('a')).must_equal [MockSearchkick::Model, 'a', {}, nil]
       args  = MockSearchkick::Model.pagy_search('a'){|a| a*2}
       block = args[-1]
-      _(args).must_equal [MockSearchkick::Model, ['a', {}], block]
+      _(args).must_equal [MockSearchkick::Model, 'a', {}, block]
     end
 
     it 'adds the caller and arguments' do
-      _(MockSearchkick::Model.pagy_search('a', b:2).results).must_equal [MockSearchkick::Model, ['a', {b: 2}], nil, :results]
-      _(MockSearchkick::Model.pagy_search('a', b:2).a('b', 2)).must_equal [MockSearchkick::Model, ['a', {b: 2}], nil, :a, 'b', 2]
+      _(MockSearchkick::Model.pagy_search('a', b:2).results).must_equal [MockSearchkick::Model, 'a', {b: 2}, nil, :results]
+      _(MockSearchkick::Model.pagy_search('a', b:2).a('b', 2)).must_equal [MockSearchkick::Model, 'a', {b: 2}, nil, :a, 'b', 2]
     end
 
   end