+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 4.0.0
 
+### Breaking Changes
+
+- renamed `Pagy::Search` as `Pagy::Searchkick` or `Pagy::ElasticsearchRails`
+
 ### Changes
 
 - Dropped support for all jruby versions and ruby version <3.0 (use pagy 3.x instead)